requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.5
boto3==1.11.14
botocore==1.14.14
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
djangorestframework==3.11.0
docutils==0.15.2
httplib2==0.9.2
idna==2.6
isort==4.3.21
jmespath==0.9.4
kazam==1.5.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.0
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.0
style==1.1.0
system-service==0.3
systemd-python==234
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
update==0.0.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.2
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
